Refactor codes to meet SOLID principle
